Verify correct presigning for Outpost ARNs
