+++ services/s3/src/test/java/software/amazon/awssdk/services/s3/S3PresignerTest.java
@@ -19,24 +19,42 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.net.URI;
+import java.time.Clock;
 import java.time.Duration;
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import org.assertj.core.data.Offset;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.AwsCredentials;
+import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.auth.signer.AwsS3V4Signer;
+import software.amazon.awssdk.auth.signer.internal.AbstractAws4Signer;
+import software.amazon.awssdk.auth.signer.internal.AbstractAwsS3V4Signer;
+import software.amazon.awssdk.auth.signer.internal.Aws4SignerRequestParams;
 import software.amazon.awssdk.auth.signer.internal.SignerConstant;
+import software.amazon.awssdk.auth.signer.params.Aws4PresignerParams;
+import software.amazon.awssdk.auth.signer.params.AwsS3V4SignerParams;
 import software.amazon.awssdk.awscore.AwsRequestOverrideConfiguration;
+import software.amazon.awssdk.core.interceptor.ExecutionAttributes;
 import software.amazon.awssdk.core.signer.NoOpSigner;
+import software.amazon.awssdk.http.SdkHttpFullRequest;
 import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.checksums.ChecksumConstant;
+import software.amazon.awssdk.services.s3.model.GetObjectRequest;
 import software.amazon.awssdk.services.s3.model.RequestPayer;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 import software.amazon.awssdk.services.s3.presigner.model.PresignedGetObjectRequest;
 import software.amazon.awssdk.services.s3.presigner.model.PresignedPutObjectRequest;
+import software.amazon.awssdk.utils.DateUtils;
 
 @RunWith(MockitoJUnitRunner.class)
 public class S3PresignerTest {
@@ -527,5 +545,111 @@ public class S3PresignerTest {
         assertThat(presignedRequest.httpRequest().rawQueryParameters().get("X-Amz-Algorithm").get(0))
             .isEqualTo("AWS4-ECDSA-P256-SHA256");
         assertThat(presignedRequest.url().toString()).startsWith(customEndpoint);
+    }
+
+    @Test
+    public void outpostArn_usWest_calculatesCorrectSignature() {
+        StaticCredentialsProvider credentials = StaticCredentialsProvider.create(AwsBasicCredentials.create(
+            "ACCESS_KEY_ID", "SECRET_ACCESS_KEY"));
+
+        String outpostArn = "arn:aws:s3-outposts:us-west-2:123456789012:outpost:op-01234567890123456:accesspoint:myaccesspoint";
+
+        S3Presigner presigner = S3Presigner.builder()
+                                           .region(Region.of("us-west-2"))
+                                           .credentialsProvider(credentials)
+                                           .serviceConfiguration(S3Configuration.builder()
+                                                                                .useArnRegionEnabled(true)
+                                                                                // Explicitly disable this because not doing so
+                                                                                // will add a new header to the signature
+                                                                                // calculation and it won't match the expected
+                                                                                // signature
+                                                                                .checksumValidationEnabled(false)
+                                                                                .build())
+                                           .build();
+
+        Duration urlDuration = Duration.ofSeconds(900);
+        ZonedDateTime signingDate = ZonedDateTime.of(2021, 8, 27, 0, 0, 0, 0, ZoneId.of("UTC"));
+        Clock signingClock = Clock.fixed(signingDate.toInstant(), ZoneId.of("UTC"));
+        Instant expirationTime = signingDate.toInstant().plus(urlDuration);
+
+        GetObjectRequest getObject = GetObjectRequest.builder()
+                                                     .bucket(outpostArn)
+                                                     .key("obj")
+                                                     .overrideConfiguration(o -> o.signer(new TestS3V4Signer(signingClock, expirationTime)))
+                                                     .build();
+
+        PresignedGetObjectRequest presigned = presigner.presignGetObject(r -> r.getObjectRequest(getObject)
+                                                                               // doesn't really do anything in this case since
+                                                                               // we set it in TestSigner
+                                                                               .signatureDuration(urlDuration));
+
+        String expectedSignature = "a944fbe2bfbae429f922746546d1c6f890649c88ba7826bd1d258ac13f327e09";
+        assertThat(presigned.url().toString()).contains("X-Amz-Signature=" + expectedSignature);
+    }
+
+    @Test
+    public void outpostArn_usEast_calculatesCorrectSignature() {
+        StaticCredentialsProvider credentials = StaticCredentialsProvider.create(AwsBasicCredentials.create(
+            "ACCESS_KEY_ID", "SECRET_ACCESS_KEY"));
+
+        String outpostArn = "arn:aws:s3-outposts:us-east-1:123456789012:outpost:op-01234567890123456:accesspoint:myaccesspoint";
+
+        S3Presigner presigner = S3Presigner.builder()
+                                           .region(Region.of("us-west-2"))
+                                           .credentialsProvider(credentials)
+                                           .serviceConfiguration(S3Configuration.builder()
+                                                                                .useArnRegionEnabled(true)
+                                                                                // Explicitly disable this because not doing so
+                                                                                // will add a new header to the signature
+                                                                                // calculation and it won't match the expected
+                                                                                // signature
+                                                                                .checksumValidationEnabled(false)
+                                                                                .build())
+                                           .build();
+
+        Duration urlDuration = Duration.ofSeconds(900);
+        ZonedDateTime signingDate = ZonedDateTime.of(2021, 8, 27, 0, 0, 0, 0, ZoneId.of("UTC"));
+        Clock signingClock = Clock.fixed(signingDate.toInstant(), ZoneId.of("UTC"));
+        Instant expirationTime = signingDate.toInstant().plus(urlDuration);
+
+        GetObjectRequest getObject = GetObjectRequest.builder()
+                                                     .bucket(outpostArn)
+                                                     .key("obj")
+                                                     .overrideConfiguration(o -> o.signer(new TestS3V4Signer(signingClock, expirationTime)))
+                                                     .build();
+
+        PresignedGetObjectRequest presigned = presigner.presignGetObject(r -> r.getObjectRequest(getObject)
+                                                                               // doesn't really do anything in this case since
+                                                                               // we set it in TestSigner
+                                                                               .signatureDuration(urlDuration));
+
+        String expectedSignature = "7f93df0b81f80e590d95442d579bd6cf749a35ff4bbdc6373fa669b89c7fce4e";
+        assertThat(presigned.url().toString()).contains("X-Amz-Signature=" + expectedSignature);
+    }
+
+    // Variant of AwsS3V4Signer that allows for changing the signing clock and expiration time
+    public static class TestS3V4Signer extends AbstractAwsS3V4Signer {
+        private final Clock signingClock;
+        private final Instant expirationTime;
+
+        public TestS3V4Signer(Clock signingClock, Instant expirationTime) {
+            this.signingClock = signingClock;
+            this.expirationTime = expirationTime;
+        }
+
+        // Intercept the presign() method so we can change the expiration
+        @Override
+        public SdkHttpFullRequest presign(SdkHttpFullRequest request, Aws4PresignerParams signingParams) {
+            Aws4PresignerParams.Builder newSignerParamsBuilder = Aws4PresignerParams.builder();
+
+            newSignerParamsBuilder.expirationTime(expirationTime);
+            newSignerParamsBuilder.signingClockOverride(signingClock);
+            newSignerParamsBuilder.signingRegion(signingParams.signingRegion());
+            newSignerParamsBuilder.signingName(signingParams.signingName());
+            newSignerParamsBuilder.awsCredentials(signingParams.awsCredentials());
+            newSignerParamsBuilder.doubleUrlEncode(signingParams.doubleUrlEncode());
+
+            return super.presign(request, newSignerParamsBuilder.build());
+        }
     }
 }
